nouveau: fix a host of warnings

+++ src/nv_mem.c
@@ -38,7 +38,7 @@ NVAllocRec *NVAllocateMemory(NVPtr pNv, int type, int size)
 				sizeof(memalloc))) {
 		ErrorF("NOUVEAU_MEM_ALLOC failed.  "
 			"flags=0x%08x, size=%lld (%d)\n",
-			memalloc.flags, memalloc.size, errno);
+			memalloc.flags, (long long)memalloc.size, errno);
 		free(mem);
 		return NULL;
 	}
@@ -48,7 +48,7 @@ NVAllocRec *NVAllocateMemory(NVPtr pNv, int type, int size)
 
 	if (drmMap(pNv->drm_fd, memalloc.map_handle, mem->size, &mem->map)) {
 		ErrorF("drmMap() failed. handle=0x%x, size=%lld (%d)\n",
-				memalloc.map_handle, mem->size, errno);
+		       memalloc.map_handle, (long long)mem->size, errno);
 		mem->map  = NULL;
 		NVFreeMemory(pNv, mem);
 		return NULL;
@@ -66,7 +66,7 @@ void NVFreeMemory(NVPtr pNv, NVAllocRec *mem)
 			if (drmUnmap(mem->map, mem->size))
 				ErrorF("drmUnmap() failed.  "
 					"map=%p, size=%lld\n",
-					mem->map, mem->size);
+					mem->map, (long long)mem->size);
 		}
 
 		memfree.flags = mem->type;
@@ -77,7 +77,7 @@ void NVFreeMemory(NVPtr pNv, NVAllocRec *mem)
 					sizeof(memfree))) {
 			ErrorF("NOUVEAU_MEM_FREE failed.  "
 				"flags=0x%08x, offset=0x%llx (%d)\n",
-				mem->type, mem->size, errno);
+				mem->type, (long long)mem->size, errno);
 		}
 		free(mem);
 	}
